fix: Resolve KSP, compilation errors, and refine notification sounds
This commit addresses build issues and completes the notification sound review.

**Build & Compilation Fixes (Phase 1):**

1.  **KSP Room Configuration:** Added `ksp { arg("room.schemaLocation", "$projectDir/schemas") }`
    to `app/build.gradle.kts` to resolve Room schema location errors for KSP.
2.  **Conflicting Imports:** Resolved ambiguous imports for `Data`,
    `ExistingWorkPolicy`, `StateFlow`, and `LocalDate` in ViewModels by
    removing duplicate import statements.
3.  **Syntax Errors:** Corrected misplaced import statements in
    `MedicationDetailHeader.kt`.
4.  **Type Mismatches & Null Safety:**
    *   Corrected argument types in `MedicationRepository.kt` (e.g., passing
        `Double?` instead of `Int` for `remainingDoses`).
    *   Ensured proper null safety for `Double?` types in
        `MedicationViewModel.kt` using `?:` operator.
5.  **Unresolved References:**
    *   Added missing `androidx.compose.runtime.remember` import in
        `MedicationCard.kt`.
    *   Added missing `com.d4viddf.medicationreminder.data.ScheduleType`
        import in `MedicationViewModel.kt`.

**Notification Sound Review (Phase 2):**

1.  **NotificationHelper Update:**
    *   Reviewed `NotificationHelper.kt`. While `RingtoneManager.TYPE_ALARM`
        was already in use, the `AudioAttributes` for the main reminder
        notification channel (`REMINDER_CHANNEL_ID`) were updated.
    *   Changed from `Notification.AUDIO_ATTRIBUTES_DEFAULT` to a more
        specific `AudioAttributes.Builder().setUsage(AudioAttributes.USAGE_ALARM).build()`.
        This ensures the notification sound is treated as an alarm by the
        system, which can affect its playback priority and behavior with
        Do Not Disturb settings.

All previously reported compilation errors should now be resolved, and the notification sound handling is more robust.

+++ app/src/main/java/com/d4viddf/medicationreminder/repository/MedicationRepository.kt
@@ -65,7 +65,8 @@ class MedicationRepository @Inject constructor(
     suspend fun updateDoseCount(medicationId: Int, remainingDoses: Int) {
         val medication = medicationDao.getMedicationById(medicationId)
         if (medication != null) {
-            val updatedMedication = medication.copy(remainingDoses = remainingDoses)
+            // Convert Int to Double for the remainingDoses field which is now Double?
+            val updatedMedication = medication.copy(remainingDoses = remainingDoses.toDouble())
             updateMedication(updatedMedication)
         }
     }
